Include ecto as a optional dependency
Geo.PostGIS.Geometry relies checks if Ecto.Type is present and if so compiles
the struct.  Using on nixos using mix2nix fails since the dependency is built
isolated and ecto is not included detected as a dependency.

Making it optional allows packaging tools to pickup the dependency and include in the
build output