Clarify compatibility issue with string arguments
We got bitten by #110 also.

The changelog is far from being clear about this change. This PR attempts to improve the situation. I'm not sure if it is better to note this in this changelog or `geo`'s or both.

Please realize that I (and probably others) consider this a breaking change. 

1) It is great that this project follows semantic versioning. In the future, consider a major bump version.

2) Consider deprecating support before breaking it

3) Consider raising a more instructive error. We currently get a `argument error` without additional detail, deep in the stack.