update geometry.ex to force recompilation on upgrade (#174)
This will fix the upgrade weirdness when switching between branches on 3.4.3 and 3.4.4.  Right now since all we did was add the optional dependency, it apparently isn't enough to encourage Elixir 1.15 to recompile this module -- however, if we actually make a (very minor) change on the upgrade, it will recompile every time.