Support Geo v4.0.0
This adds support for either releases of Geo compatible with v3.6 *or* v4.0. For our purposes, the breaking changes in v4.0.0 are not signficant.

I've tested overriding the Geo dependency to v4.0.0 on the Felt codebase, where we have a great deal of unit tests round-tripping imports of GeoJSON, and aside from the default SRID change, everything Just Works™.